+++ src/main/java/dev/nullpanic/tamagotchitelegrambot/service/PetService.java
@@ -1,0 +1,8 @@
+package dev.nullpanic.tamagotchitelegrambot.service;
+
+import dev.nullpanic.tamagotchitelegrambot.persist.model.Pet;
+
+public interface PetService {
+
+    Pet createPet(Pet Pet);
+}